[fix] migrateできない問題を修正

+++ api/db/migrate/20260605063647_add_uses_place_id_to_groups.rb
@@ -1,6 +1,6 @@
 class AddUsesPlaceIdToGroups < ActiveRecord::Migration[6.1]
   def change
-    add_column :groups, :uses_place_id, :integer, null: true
+    add_column :groups, :uses_place_id, :bigint, null: true
     add_index :groups, :uses_place_id
     add_foreign_key :groups, :stocker_places, column: :uses_place_id
   end
